backend/requirements.txt
bcrypt==4.0.1
click==8.1.3
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
psycopg2-binary==2.9.5
PyMySQL==1.0.2
python-decouple==3.8
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
